+++ lib/src/http/angel_base.dart
@@ -30,7 +30,8 @@ class AngelBase extends Routable {
   /// A function that renders views.
   ///
   /// Called by [ResponseContext]@`render`.
-  ViewGenerator viewGenerator = (String view, [Map data]) async => "No view engine has been configured yet.";
+  ViewGenerator viewGenerator = (String view, [Map data]) async =>
+      "No view engine has been configured yet.";
 
   operator [](key) => properties[key];
   operator []=(key, value) => properties[key] = value;